fix: 탈퇴로직 변경
